refactor instagram likes page components
