+++ cicero-dashboard/components/likes/instagram/Divider.jsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function Divider() {
+  return <div className="hidden md:block w-px bg-gray-200 mx-2 my-2"></div>;
+}
